- naliczenia i skladniki

+++ src/main/resources/scripts/sources/entities/element.py
@@ -22,7 +22,7 @@ class ElementManager(Container):
     def setElementData(self, element):
         element.setName(vars.get(self._prefix + 'name'))
         element.setKey(vars.get(self._prefix + 'key'))
-        element.setGlobalValue(double(vars.get(self._prefix + 'globalValue')))
+        element.setGlobalValue(float(vars.get(self._prefix + 'globalValue')))
         element.setGroup(self.findGroupByKey(vars.get(self._prefix + 'group')))
         
     def saveElement(self, element):
